Revert "Merge branch 'master' of https://github.com/space-wizards/space-station-14 into headsets"
This reverts commit 463a2839e22c793889a95b89a5e9b25347d66e0a, reversing
changes made to efa4729f17a115dda2faa97255925909ea01cec5.

probably did this wrong too, reverting just in case

+++ Content.Server/Atmos/Hotspot.cs
@@ -23,7 +23,7 @@ namespace Content.Server.Atmos
         ///     State for the fire sprite.
         /// </summary>
         [ViewVariables]
-        public byte State;
+        public int State;
 
         public void Start()
         {
